backend/requirements.txt
fastapi
uvicorn
pydantic
mistralai
python-dotenv